test: Separate module tests from fixture-driven tests
Fixtures live only in test/fixtures, and a test that consumes one is
not a unit test: the blueprint-driven suites (commands/spec,
commands/registry, render/help, render/completion) move under test/,
mirroring the source layout, along with every test that reaches into
another module for its fakes (context, auth/operations, http/request,
interact-for-blueprint-object).

A test now sits beside its module in src only when it tests the module
of the same name and imports nothing beyond it, external packages and
type-only imports excepted. TESTING.md records the rule.
