+++ test/render/help.test.ts
@@ -1,8 +1,9 @@
 import { expect, test } from 'vitest'
 
-import { testBlueprint } from '../../../test/fixtures/blueprint.js'
-import { buildRegistry } from '../commands/registry.js'
-import { renderHelp } from '../render/help.js'
+import { buildRegistry } from 'lib/commands/registry.js'
+import { renderHelp } from 'lib/render/help.js'
+
+import { testBlueprint } from '../fixtures/blueprint.js'
 
 const { spec } = buildRegistry(testBlueprint)
